jsonpath_rust::parser

Module model

Structs

JpQuery

Represents a JSONPath query with a list of segments.

Enums

Comparable

Enum representing different types of comparable values in a JSONPath query.

Comparison

Enum representing different types of comparisons in a JSONPath query.

Filter

Enum representing different types of filters in a JSONPath query.

FilterAtom

Enum representing different types of atomic filters in a JSONPath query.

FnArg

Enum representing different types of function arguments in a JSONPath query.

Literal

Enum representing different types of literal values in a JSONPath query.

Segment

Enum representing different types of segments in a JSONPath query.

Selector

Enum representing different types of selectors in a JSONPath query.

SingularQuery

Enum representing different types of singular queries in a JSONPath query.

SingularQuerySegment

Enum representing different types of singular query segments in a JSONPath query.

Test

Enum representing different types of tests in a JSONPath query.

TestFunction

Enum representing different types of test functions in a JSONPath query.

Functions

slice_from