jsonpath_rust::path

Trait JsonLike

pub trait JsonLike:
    Default
    + Clone
    + Debug
    + for<'a> From<&'a str>
    + From<Vec<String>>
    + From<bool>
    + From<i64>
    + From<f64>
    + From<Vec<Self>>
    + From<String>
    + PartialEq
    + 'static {

    // Required methods
    fn get(&self, key: &str) -> Option<&Self>;
    fn itre(&self, pref: String) -> Vec<JsonPathValue<'_, Self>>;
    fn array_len(&self) -> JsonPathValue<'static, Self>;
    fn init_with_usize(cnt: usize) -> Self;
    fn deep_flatten(&self, pref: String) -> Vec<(&Self, String)>;
    fn deep_path_by_key<'a>(
        &'a self,
        key: ObjectField<'a, Self>,
        pref: String,
    ) -> Vec<(&'a Self, String)>;
    fn as_u64(&self) -> Option<u64>;
    fn is_array(&self) -> bool;
    fn as_array(&self) -> Option<&Vec<Self>>;
    fn size(left: Vec<&Self>, right: Vec<&Self>) -> bool;
    fn sub_set_of(left: Vec<&Self>, right: Vec<&Self>) -> bool;
    fn any_of(left: Vec<&Self>, right: Vec<&Self>) -> bool;
    fn regex(left: Vec<&Self>, right: Vec<&Self>) -> bool;
    fn inside(left: Vec<&Self>, right: Vec<&Self>) -> bool;
    fn less(left: Vec<&Self>, right: Vec<&Self>) -> bool;
    fn eq(left: Vec<&Self>, right: Vec<&Self>) -> bool;
    fn null() -> Self;
    fn array(data: Vec<Self>) -> Self;
    fn reference<T>(
        &self,
        path: T,
    ) -> Result<Option<&Self>, JsonPathParserError>
       where T: Into<JsonPathStr>;
    fn reference_mut<T>(
        &mut self,
        path: T,
    ) -> Result<Option<&mut Self>, JsonPathParserError>
       where T: Into<JsonPathStr>;

}

The JsonLike trait defines a set of methods and associated types for working with JSON-like data structures.

It provides a common interface for accessing and manipulating JSON data, allowing for operations such as

retrieving values by key,

iterating over elements

performing various comparisons and transformations.

The trait is implemented for the serde_json::Value type already

Required Methods

fn get(&self, key: &str) -> Option<&Self>

Retrieves a reference to the value associated with the given key.

fn itre(&self, pref: String) -> Vec<JsonPathValue<'_, Self>>

Iterates over the elements with a given prefix and returns a vector of JsonPathValue.

fn array_len(&self) -> JsonPathValue<'static, Self>

Returns the length of the array as a JsonPathValue.

fn init_with_usize(cnt: usize) -> Self

Initializes an instance with a specific size.

fn deep_flatten(&self, pref: String) -> Vec<(&Self, String)>

Flattens nested structures and returns a vector of tuples containing references to the elements and their paths.

fn deep_path_by_key<'a>( &'a self, key: ObjectField<'a, Self>, pref: String, ) -> Vec<(&'a Self, String)>

Performs a deep search by key and returns a vector of tuples containing references to the elements and their paths.

fn as_u64(&self) -> Option<u64>

Converts the element to an Option<u64>.

fn is_array(&self) -> bool

Checks if the element is an array.

fn as_array(&self) -> Option<&Vec<Self>>

Converts the element to an Option<&Vec<Self>>.

fn size(left: Vec<&Self>, right: Vec<&Self>) -> bool

Compares the size of two vectors of references to elements.

fn sub_set_of(left: Vec<&Self>, right: Vec<&Self>) -> bool

Checks if the left vector is a subset of the right vector.

fn any_of(left: Vec<&Self>, right: Vec<&Self>) -> bool

Checks if any element in the left vector is present in the right vector.

fn regex(left: Vec<&Self>, right: Vec<&Self>) -> bool

Checks if the elements in the left vector match the regex pattern in the right vector.

fn inside(left: Vec<&Self>, right: Vec<&Self>) -> bool

Checks if any element in the left vector is inside the right vector.

fn less(left: Vec<&Self>, right: Vec<&Self>) -> bool

Ensures the number on the left side is less than the number on the right side.

fn eq(left: Vec<&Self>, right: Vec<&Self>) -> bool

Compares elements for equality.

fn null() -> Self

Returns a null value.

fn array(data: Vec<Self>) -> Self

Creates an array from a vector of elements.

fn reference<T>(&self, path: T) -> Result<Option<&Self>, JsonPathParserError>
where T: Into<JsonPathStr>,

Retrieves a reference to the element at the specified path. The path is specified as a string and can be obtained from the query.

Arguments

path - A json path to the element specified as a string (root, field, index only).

fn reference_mut<T>( &mut self, path: T, ) -> Result<Option<&mut Self>, JsonPathParserError>
where T: Into<JsonPathStr>,

Retrieves a mutable reference to the element at the specified path.

Arguments

path - A json path to the element specified as a string (root, field, index only).

Examples

use serde_json::json;
use jsonpath_rust::{JsonPath, JsonPathParserError};
use jsonpath_rust::path::JsonLike;

let mut json = json!([
    {"verb": "RUN","distance":[1]},
    {"verb": "TEST"},
    {"verb": "DO NOT RUN"}
]);

let path: Box<JsonPath> = Box::from(JsonPath::try_from("$.[?(@.verb == 'RUN')]").unwrap());
let elem = path
    .find_as_path(&json)
    .get(0)
    .cloned()
    .ok_or(JsonPathParserError::InvalidJsonPath("".to_string())).unwrap();

if let Some(v) = json
    .reference_mut(elem).unwrap()
    .and_then(|v| v.as_object_mut())
    .and_then(|v| v.get_mut("distance"))
    .and_then(|v| v.as_array_mut())
{
    v.push(json!(2))
}

assert_eq!(
    json,
    json!([
        {"verb": "RUN","distance":[1,2]},
        {"verb": "TEST"},
        {"verb": "DO NOT RUN"}
    ])
);

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.