jsonpath_rust

Module query

Modules

queryable

state

Structs

QueryRef

The resulting type of JSONPath query. It can either be a value or a reference to a value with its path.

Traits

Query

Main internal trait to implement the logic of processing jsonpath.

Functions

js_path

The main function to process a JSONPath query. It takes a path and a value, and returns a vector of QueryResult thus values + paths.

js_path_path

A convenience function to process a JSONPath query and return a vector of paths, omitting the values.

js_path_process

A convenience function to process a JSONPath query and return a vector of values, omitting the path.

js_path_vals

A convenience function to process a JSONPath query and return a vector of values, omitting the path.

Type Aliases

Queried

A type that can be queried with JSONPath, typically Result

QueryPath

A type that can be queried with JSONPath, typically string