Enum jsonpath_rust::JsonPtr

pub enum JsonPtr<'a, Data> {
    Slice(&'a Data),
    NewValue(Data),
}

Json paths may return either pointers to the original json or new data. This custom pointer type allows us to handle both cases. Unlike JsonPathValue, this type does not represent NoValue to allow the implementation of Deref.

Variants

Slice(&'a Data)

The slice of the initial json data

NewValue(Data)

The new data that was generated from the input data (like length operator)