Module jsonpath_rust::path

Traits

Path

The trait defining the behaviour of processing every separated element. type Data usually stands for json [Value] The trait also requires to have a root json to process. It needs in case if in the filter there will be a pointer to the absolute path

Functions

json_path_instance

The major method to process the top part of json part

Type Aliases

PathInstance

The basic type for instances.