Struct jsonpath_rust::JsonPathFinder

pub struct JsonPathFinder { /* private fields */ }

The base structure stitching the json instance and jsonpath instance

Implementations

impl JsonPathFinder

pub fn new(json: Box<Value>, path: Box<JsonPathInst>) -> Self

creates a new instance of JsonPathFinder

pub fn set_path(&mut self, path: Box<JsonPathInst>)

updates a path with a new one

pub fn set_json(&mut self, json: Box<Value>)

updates a json with a new one

pub fn set_json_str(&mut self, json: &str) -> Result<(), String>

updates a json from string and therefore can be some parsing errors

pub fn set_path_str(&mut self, path: &str) -> Result<(), String>

updates a path from string and therefore can be some parsing errors

pub fn from_str(json: &str, path: &str) -> Result<Self, String>

create a new instance from string and therefore can be some parsing errors

pub fn instance(&self) -> PathInstance<'_>

creates an instance to find a json slice from the json

pub fn find_slice(&self) -> Vec<JsonPathValue<'_, Value>>

finds a slice of data in the set json. The result is a vector of references to the incoming structure.

pub fn find(&self) -> Value

finds a slice of data and wrap it with Value::Array by cloning the data. Returns either an array of elements or Json::Null if the match is incorrect.

pub fn find_as_path(&self) -> Value

finds a path of the values. If the values has been obtained by moving the data out of the initial json the path is absent.